requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.1.0
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.3
django-celery-beat==2.2.0
django-filter==2.4.0
djangorestframework==3.12.4
docker==5.0.0
gunicorn==20.1.0
humanize==3.7.0
idna==2.10
kombu==5.1.0
Markdown==3.3.4
prometheus-client==0.8.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
pytz==2021.1
redis==3.5.3
requests==2.25.1
selenium==3.141.0
six==1.16.0
sqlparse==0.4.1
tornado==6.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.0.1
django-cors-headers==3.7.0
pyjwt==2.1.0
djangorestframework-simplejwt==4.7.2
stripe==2.60.0
twilio==7.3.0
slack-sdk==3.11.2
django-markdownify==0.9.0